requirements.txt
# Core data
yfinance>=0.2.36,<0.3
pandas>=2.0.0,<3.0
numpy>=1.24.0,<2.0
scipy>=1.11.0,<2.0

# Brokers
alpaca-py>=0.8.2,<1.0
python-binance>=1.0.19,<2.0

# LLM Judge
anthropic>=0.28.0,<1.0

# Utilities
python-dotenv>=1.0.0,<2.0
requests>=2.31.0,<3.0
colorama>=0.4.6,<1.0
tabulate>=0.9.0,<1.0
